refactor: added dept role components
