+++ client/src/components/admin/departments/AdminDeptUserView.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { Header, Grid, Breadcrumb, Card, Image } from 'semantic-ui-react';
+import { Header, Grid, Breadcrumb, Card } from 'semantic-ui-react';
 import { isEmpty } from 'lodash';
 
-import { roleUsers, noRoleUsers } from '../../../utils/roleHelpers';
+import { noRoleUsers } from '../../../utils/roleHelpers';
 
 import {
   selectUsersInDept,
@@ -12,61 +12,22 @@ import {
 } from '../../../reducers/selectors/adminSelectors';
 import { selectDept } from '../../../reducers/selectors/deptSelectors';
 
+import AdminDeptRoleUsers from './AdminDeptRoleUsers';
+import AdminDeptNoRoleUsers from './AdminDeptNoRoleUsers';
+
 class AdminDeptUserView extends Component {
   renderRoleColumns() {
     const { roles } = this.props;
     return roles.map((role) => {
       return (
         <Grid.Column key={role._id}>
-          <Header as="h4" textAlign="center">
+          <Header as="h5" textAlign="center">
             {role.roleName}
           </Header>
           <Card.Group itemsPerRow={3} centered>
-            {this.renderRoleUsers(role._id)}
+            <AdminDeptRoleUsers roleId={role._id} />
           </Card.Group>
         </Grid.Column>
-      );
-    });
-  }
-
-  renderRoleUsers(roleId) {
-    const { deptUsers } = this.props;
-    const rUsers = roleUsers(deptUsers, roleId);
-    return rUsers.map((user, index) => {
-      return (
-        <Card key={index} fluid>
-          <Card.Content>
-            <Image floated="right" size="mini" src={user.imageUrl} />
-            <Card.Header>
-              {user.firstName} {user.lastName}
-            </Card.Header>
-            <Card.Meta>{user.username}</Card.Meta>
-            <Card.Description>
-              Reserved for adding in Role competencies here (and status)
-            </Card.Description>
-          </Card.Content>
-        </Card>
-      );
-    });
-  }
-
-  renderNoRoleUsers() {
-    const { deptUsers } = this.props;
-    const nrUsers = noRoleUsers(deptUsers);
-    return nrUsers.map((user, index) => {
-      return (
-        <Card key={index} fluid>
-          <Card.Content>
-            <Image floated="right" size="mini" src={user.imageUrl} />
-            <Card.Header>
-              {user.firstName} {user.lastName}
-            </Card.Header>
-            <Card.Meta>{user.username}</Card.Meta>
-            <Card.Description>
-              Reserved for adding in Role competencies here (and status)
-            </Card.Description>
-          </Card.Content>
-        </Card>
       );
     });
   }
@@ -98,6 +59,13 @@ class AdminDeptUserView extends Component {
           <Grid celled centered>
             {roles.length > 0 && (
               <Grid.Row columns={roles.length}>
+                <Header
+                  as="h4"
+                  textAlign="center"
+                  style={{ marginTop: '0.5em' }}
+                >
+                  Roles
+                </Header>
                 {this.renderRoleColumns()}
               </Grid.Row>
             )}
@@ -114,7 +82,8 @@ class AdminDeptUserView extends Component {
                 Users with no Roles
               </Header>
               <Card.Group itemsPerRow={4} centered>
-                {this.renderNoRoleUsers()}
+                {/* {this.renderNoRoleUsers()} */}
+                <AdminDeptNoRoleUsers />
               </Card.Group>
             </Grid.Column>
           </Grid>
